requirements.txt
rapidfuzz
sqlalchemy
fastapi
uvicorn[standard]
pydantic
openai
tqdm
jinja2
python-multipart
python-dotenv